requirements.txt
asyncpg~=0.28.0
aiogram~=3.0.0rc2
python-dotenv~=1.0.0
redis~=5.0.0
loguru~=0.7.0